--- V1.3/art.py
@@ -1,17 +1,0 @@
-logo = '''
-
-  _   _ ______  __      _       _                               _    _                    _    ____        _                                      
- | \ | / ___\ \/ /     / \   __| |_   ____ _ _ __   ___ ___  __| |  | |    ___   __ _  __| |  | __ )  __ _| | __ _ _ __   ___ ___ _ __            
- |  \| \___ \\  /     / _ \ / _` \ \ / / _` | '_ \ / __/ _ \/ _` |  | |   / _ \ / _` |/ _` |  |  _ \ / _` | |/ _` | '_ \ / __/ _ \ '__|           
- | |\  |___) /  \    / ___ \ (_| |\ V / (_| | | | | (_|  __/ (_| |  | |__| (_) | (_| | (_| |  | |_) | (_| | | (_| | | | | (_|  __/ |              
- |_| \_|____/_/\_\  /_/   \_\__,_| \_/ \__,_|_| |_|\___\___|\__,_|  |_____\___/ \__,_|\__,_|  |____/ \__,_|_|\__,_|_| |_|\___\___|_|              
-  __     ___      _               _    ____                  _             __  __ _                 _                        _   ____             
-  \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___   |  \/  (_) __ _ _ __ __ _| |_ ___  _ __   __   __ / | |___ \            
-   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \  | |\/| | |/ _` | '__/ _` | __/ _ \| '__|  \ \ / / | |   __) |           
-    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/  | |  | | | (_| | | | (_| | || (_) | |      \ V /  | |_ / __/            
-     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___|  |_|  |_|_|\__, |_|  \__,_|\__\___/|_|       \_/   |_(_)_____|           
-                                                                                    |___/                                                         
-                                      
-'''
-line = ''' ####################################################################################################################################################
-'''
