--- V1.3/titles.py
@@ -1,215 +1,0 @@
-disclaimer = '''
-INTENDED ONLY FOR NON-PRODUCTION AND HOME LABS
-'''
-
-login = '''
-  _                _                          _             _       ____ ____  ____  _____ _____     _              
- | |    ___   __ _(_)_ __      __ _ _ __   __| |   ___  ___| |_    / ___/ ___||  _ \|  ___|_   _|__ | | _____ _ __  
- | |   / _ \ / _` | | '_ \    / _` | '_ \ / _` |  / __|/ _ \ __|  | |   \___ \| |_) | |_    | |/ _ \| |/ / _ \ '_ \ 
- | |__| (_) | (_| | | | | |  | (_| | | | | (_| |  \__ \  __/ |_   | |___ ___) |  _ <|  _|   | | (_) |   <  __/ | | |
- |_____\___/ \__, |_|_| |_|   \__,_|_| |_|\__,_|  |___/\___|\__|   \____|____/|_| \_\_|     |_|\___/|_|\_\___|_| |_|
-             |___/                                                                                                  
-'''
-
-cloud = '''
-
-  ____       _           _      _____                    _      _   _ ______  __      _    _     ____      ____ _                 _ 
- / ___|  ___| | ___  ___| |_   |_   _|_ _ _ __ __ _  ___| |_   | \ | / ___\ \/ /     / \  | |   | __ )    / ___| | ___  _   _  __| |
- \___ \ / _ \ |/ _ \/ __| __|    | |/ _` | '__/ _` |/ _ \ __|  |  \| \___ \\  /     / _ \ | |   |  _ \   | |   | |/ _ \| | | |/ _` |
-  ___) |  __/ |  __/ (__| |_     | | (_| | | | (_| |  __/ |_   | |\  |___) /  \    / ___ \| |___| |_) |  | |___| | (_) | |_| | (_| |
- |____/ \___|_|\___|\___|\__|    |_|\__,_|_|  \__, |\___|\__|  |_| \_|____/_/\_\  /_/   \_\_____|____/    \____|_|\___/ \__,_|\__,_|
-                                              |___/                                                                                 
-
-'''
-
-dnsprofile = '''
-
-  ____                     ____  _   _ ____     ____             __ _ _           
- / ___|  ___ __ _ _ __    |  _ \| \ | / ___|   |  _ \ _ __ ___  / _(_) | ___  ___ 
- \___ \ / __/ _` | '_ \   | | | |  \| \___ \   | |_) | '__/ _ \| |_| | |/ _ \/ __|
-  ___) | (_| (_| | | | |  | |_| | |\  |___) |  |  __/| | | (_) |  _| | |  __/\__ \.
- |____/ \___\__,_|_| |_|  |____/|_| \_|____/   |_|   |_|  \___/|_| |_|_|\___||___/
-                                                                                  
-'''
-
-ipamprofile = '''
-
-  ____                     ___ ____   _    __  __    ____             __ _ _           
- / ___|  ___ __ _ _ __    |_ _|  _ \ / \  |  \/  |  |  _ \ _ __ ___  / _(_) | ___  ___ 
- \___ \ / __/ _` | '_ \    | || |_) / _ \ | |\/| |  | |_) | '__/ _ \| |_| | |/ _ \/ __|
-  ___) | (_| (_| | | | |   | ||  __/ ___ \| |  | |  |  __/| | | (_) |  _| | |  __/\__ \.
- |____/ \___\__,_|_| |_|  |___|_| /_/   \_\_|  |_|  |_|   |_|  \___/|_| |_|_|\___||___/
-                                                                                       
-'''
-
-vrfcontext = '''
-
-  ____       _           _      _____                    _     __     ______  _____     ____            _            _   
- / ___|  ___| | ___  ___| |_   |_   _|_ _ _ __ __ _  ___| |_   \ \   / /  _ \|  ___|   / ___|___  _ __ | |_ _____  _| |_ 
- \___ \ / _ \ |/ _ \/ __| __|    | |/ _` | '__/ _` |/ _ \ __|   \ \ / /| |_) | |_     | |   / _ \| '_ \| __/ _ \ \/ / __|
-  ___) |  __/ |  __/ (__| |_     | | (_| | | | (_| |  __/ |_     \ V / |  _ <|  _|    | |__| (_) | | | | ||  __/>  <| |_ 
- |____/ \___|_|\___|\___|\__|    |_|\__,_|_|  \__, |\___|\__|     \_/  |_| \_\_|       \____\___/|_| |_|\__\___/_/\_\\__|
-                                              |___/                                                                      
-
-'''
-serviceenginegroup = '''
-
-  ____       _           _      _____                    _      ____                  _             _____             _             
- / ___|  ___| | ___  ___| |_   |_   _|_ _ _ __ __ _  ___| |_   / ___|  ___ _ ____   _(_) ___ ___   | ____|_ __   __ _(_)_ __   ___  
- \___ \ / _ \ |/ _ \/ __| __|    | |/ _` | '__/ _` |/ _ \ __|  \___ \ / _ \ '__\ \ / / |/ __/ _ \  |  _| | '_ \ / _` | | '_ \ / _ \ 
-  ___) |  __/ |  __/ (__| |_     | | (_| | | | (_| |  __/ |_    ___) |  __/ |   \ V /| | (_|  __/  | |___| | | | (_| | | | | |  __/ 
- |____/_\___|_|\___|\___|\__|    |_|\__,_|_|  \__, |\___|\__|  |____/ \___|_|    \_/ |_|\___\___|  |_____|_| |_|\__, |_|_| |_|\___| 
-   / ___|_ __ ___  _   _ _ __                 |___/                                                             |___/               
-  | |  _| '__/ _ \| | | | '_ \                                                                                                      
-  | |_| | | | (_) | |_| | |_) |                                                                                                     
-   \____|_|  \___/ \__,_| .__/                                                                                                      
-                        |_|                                                                                                         
-
-'''
-
-vs_selector = '''
-
-  ____       _           _     __     ___      _               _    ____                  _                
- / ___|  ___| | ___  ___| |_   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___  
- \___ \ / _ \ |/ _ \/ __| __|   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __| 
-  ___) |  __/ |  __/ (__| |_     \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \ 
- |____/ \___|_|\___|\___|\__|     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/ 
-                                                                                                           
-
-'''
-
-httppolicyset_selector = '''
-
-  ____       _           _      _   _ _____ _____ ____     ____       _ _               ____       _          
- / ___|  ___| | ___  ___| |_   | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___    
- \___ \ / _ \ |/ _ \/ __| __|  | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|   
-  ___) |  __/ |  __/ (__| |_   |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \   
- |____/ \___|_|\___|\___|\__|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/   
-                                                                               |___/                          
-
-'''
-
-httppolicyset_scanner = '''
-
-  ____                     _   _ _____ _____ ____     ____       _ _               ____       _                   
- / ___|  ___ __ _ _ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___             
- \___ \ / __/ _` | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|            
-  ___) | (_| (_| | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \            
- |____/ \___\__,_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/            
-   __               ____             _                      ____          |___/   ____                            
-  / _| ___  _ __   |  _ \ ___   ___ | |___     ___  _ __   |  _ \ ___   ___ | |  / ___|_ __ ___  _   _ _ __  ___  
- | |_ / _ \| '__|  | |_) / _ \ / _ \| / __|   / _ \| '__|  | |_) / _ \ / _ \| | | |  _| '__/ _ \| | | | '_ \/ __| 
- |  _| (_) | |     |  __/ (_) | (_) | \__ \  | (_) | |     |  __/ (_) | (_) | | | |_| | | | (_) | |_| | |_) \__ \ 
- |_|  \___/|_|     |_|   \___/ \___/|_|___/   \___/|_|     |_|   \___/ \___/|_|  \____|_|  \___/ \__,_| .__/|___/ 
-                                                                                                      |_|         
-
-'''
-
-httppolicyset_migrate_pools = '''
-
-  __  __ _                 _          ____             _        _          _   _ _____ _____ ____     ____       _ _               ____       _           
- |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |___   (_)_ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
- | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| / __|  | | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
- | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | \__ \  | | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
- |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|___/  |_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
-           |___/                                                                                                          |___/                           
-
-'''
-
-httppolicyset_migratepoolgroups = '''
-
-  __  __ _                 _          ____             _     ____                              _          _   _ _____ _____ ____     ____       _ _               ____       _           
- |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |   / ___|_ __ ___  _   _ _ __  ___   (_)_ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
- | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| |  | |  _| '__/ _ \| | | | '_ \/ __|  | | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
- | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | |  | |_| | | | (_) | |_| | |_) \__ \  | | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
- |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|   \____|_|  \___/ \__,_| .__/|___/  |_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
-           |___/                                                                 |_|                                                                     |___/                           
-
-'''
-
-httppolicyset_migrate = '''
-
-  __  __ _                 _          _   _ _____ _____ ____     ____       _ _               ____       _           
- |  \/  (_) __ _ _ __ __ _| |_ ___   | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
- | |\/| | |/ _` | '__/ _` | __/ _ \  | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
- | |  | | | (_| | | | (_| | ||  __/  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
- |_|  |_|_|\__, |_|  \__,_|\__\___|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
-           |___/                                                                     |___/                           
-
-'''
-
-vs_migratepoolgroups = '''
-
-  __  __ _                 _          ____             _     ____                                     __   __     ___      _               _    ____                  _                  
- |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |   / ___|_ __ ___  _   _ _ __  ___     ___  / _|  \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
- | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| |  | |  _| '__/ _ \| | | | '_ \/ __|   / _ \| |_    \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
- | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | |  | |_| | | | (_) | |_| | |_) \__ \  | (_) |  _|    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
- |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|   \____|_|  \___/ \__,_| .__/|___/   \___/|_|       \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
-           |___/                                                                 |_|                                                                                                     
-
-'''
-
-vs_migratepools = '''
-
-  __  __ _                 _          ____             _               __   __     ___      _               _    ____                  _                  
- |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |___     ___  / _|  \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
- | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| / __|   / _ \| |_    \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
- | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | \__ \  | (_) |  _|    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
- |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|___/   \___/|_|       \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
-           |___/                                                                                                                                          
-
-'''
-
-vsvip_migrate = '''
-
-  __  __ _                 _         __     ______   __     _____ ____       
- |  \/  (_) __ _ _ __ __ _| |_ ___   \ \   / / ___|  \ \   / /_ _|  _ \ ___  
- | |\/| | |/ _` | '__/ _` | __/ _ \   \ \ / /\___ \   \ \ / / | || |_) / __| 
- | |  | | | (_| | | | (_| | ||  __/    \ V /  ___) |   \ V /  | ||  __/\__ \ 
- |_|  |_|_|\__, |_|  \__,_|\__\___|     \_/  |____/     \_/  |___|_|   |___/ 
-           |___/                                                             
-
-'''
-
-vs_migrate = '''
-
-  __  __ _                 _         __     ___      _               _    ____                  _                  
- |  \/  (_) __ _ _ __ __ _| |_ ___   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
- | |\/| | |/ _` | '__/ _` | __/ _ \   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
- | |  | | | (_| | | | (_| | ||  __/    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
- |_|  |_|_|\__, |_|  \__,_|\__\___|     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
-           |___/                                                                                                   
-
-'''
-
-logout = '''
-
-  _                            _                      _                   _                      _             
- | |    ___   __ _  ___  _   _| |_     __ _ _ __   __| |    ___ _ __   __| |   ___  ___  ___ ___(_) ___  _ __  
- | |   / _ \ / _` |/ _ \| | | | __|   / _` | '_ \ / _` |   / _ \ '_ \ / _` |  / __|/ _ \/ __/ __| |/ _ \| '_ \ 
- | |__| (_) | (_| | (_) | |_| | |_   | (_| | | | | (_| |  |  __/ | | | (_| |  \__ \  __/\__ \__ \ | (_) | | | |
- |_____\___/ \__, |\___/ \__,_|\__|   \__,_|_| |_|\__,_|   \___|_| |_|\__,_|  |___/\___||___/___/_|\___/|_| |_|
-             |___/                                                                                             
-
-'''
-
-thanks = '''
-
-  _____ _                 _            __                         _                                                                                    
- |_   _| |__   __ _ _ __ | | _____    / _| ___  _ __    _   _ ___(_)_ __   __ _                                                                        
-   | | | '_ \ / _` | '_ \| |/ / __|  | |_ / _ \| '__|  | | | / __| | '_ \ / _` |                                                                       
-   | | | | | | (_| | | | |   <\__ \  |  _| (_) | |     | |_| \__ \ | | | | (_| |                                                                       
-   |_| |_| |_|\__,_|_| |_|_|\_\___/  |_|  \___/|_|      \__,_|___/_|_| |_|\__, |                                                                       
-   _   _ ______  __      _       _                               _    _   |___/            _    ____        _                                          
-  | \ | / ___\ \/ /     / \   __| |_   ____ _ _ __   ___ ___  __| |  | |    ___   __ _  __| |  | __ )  __ _| | __ _ _ __   ___ ___ _ __                
-  |  \| \___ \\  /     / _ \ / _` \ \ / / _` | '_ \ / __/ _ \/ _` |  | |   / _ \ / _` |/ _` |  |  _ \ / _` | |/ _` | '_ \ / __/ _ \ '__|               
-  | |\  |___) /  \    / ___ \ (_| |\ V / (_| | | | | (_|  __/ (_| |  | |__| (_) | (_| | (_| |  | |_) | (_| | | (_| | | | | (_|  __/ |                  
-  |_| \_|____/_/\_\  /_/   \_\__,_| \_/ \__,_|_| |_|\___\___|\__,_|  |_____\___/ \__,_|\__,_|  |____/ \__,_|_|\__,_|_| |_|\___\___|_|                  
-   __     ___      _               _    ____                  _             __  __ _                 _                        _   ____                 
-   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___   |  \/  (_) __ _ _ __ __ _| |_ ___  _ __   __   __ / | |___ \                
-    \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \  | |\/| | |/ _` | '__/ _` | __/ _ \| '__|  \ \ / / | |   __) |               
-     \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/  | |  | | | (_| | | | (_| | || (_) | |      \ V /  | |_ / __/                
-      \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___|  |_|  |_|_|\__, |_|  \__,_|\__\___/|_|       \_/   |_(_)_____|               
-                                                                                     |___/                                                             
-
-'''
