+++ V1.4/class_modules/titles.py
@@ -1,0 +1,265 @@
+titles_disclaimer = '''
+INTENDED ONLY FOR NON-PRODUCTION AND HOME LABS
+'''
+
+titles_logo = '''
+
+                          __     __ __  __                                    _ __     __ ___                                           
+                          \ \   / /|  \/  |__      __ __ _  _ __  ___        / \\ \   / /|_ _|                                          
+                           \ \ / / | |\/| |\ \ /\ / // _` || '__|/ _ \      / _ \\ \ / /  | |                                           
+                            \ V /  | |  | | \ V  V /| (_| || |  |  __/     / ___ \\ V /   | |                                           
+                             \_/   |_|  |_|  \_/\_/  \__,_||_|   \___|    /_/   \_\\_/   |___|                                          
+ __     __ _        _                  _      ____                      _                 __  __  _                     _               
+ \ \   / /(_) _ __ | |_  _   _   __ _ | |    / ___|   ___  _ __ __   __(_)  ___  ___     |  \/  |(_)  __ _  _ __  __ _ | |_  ___   _ __ 
+  \ \ / / | || '__|| __|| | | | / _` || |    \___ \  / _ \| '__|\ \ / /| | / __|/ _ \    | |\/| || | / _` || '__|/ _` || __|/ _ \ | '__|
+   \ V /  | || |   | |_ | |_| || (_| || |     ___) ||  __/| |    \ V / | || (__|  __/    | |  | || || (_| || |  | (_| || |_| (_) || |   
+    \_/   |_||_|    \__| \__,_| \__,_||_|    |____/  \___||_|     \_/  |_| \___|\___|    |_|  |_||_| \__, ||_|   \__,_| \__|\___/ |_|   
+                                                                                                     |___/                              
+
+'''
+
+titles_login = '''
+  _                _                          _             _       ____ ____  ____  _____ _____     _              
+ | |    ___   __ _(_)_ __      __ _ _ __   __| |   ___  ___| |_    / ___/ ___||  _ \|  ___|_   _|__ | | _____ _ __  
+ | |   / _ \ / _` | | '_ \    / _` | '_ \ / _` |  / __|/ _ \ __|  | |   \___ \| |_) | |_    | |/ _ \| |/ / _ \ '_ \ 
+ | |__| (_) | (_| | | | | |  | (_| | | | | (_| |  \__ \  __/ |_   | |___ ___) |  _ <|  _|   | | (_) |   <  __/ | | |
+ |_____\___/ \__, |_|_| |_|   \__,_|_| |_|\__,_|  |___/\___|\__|   \____|____/|_| \_\_|     |_|\___/|_|\_\___|_| |_|
+             |___/                                                                                                  
+'''
+
+titles_cloud = '''
+
+  ____       _      _____                    _         ___     _____     ____ _                 _ 
+ / ___|  ___| |_   |_   _|_ _ _ __ __ _  ___| |_      / \ \   / /_ _|   / ___| | ___  _   _  __| |
+ \___ \ / _ \ __|    | |/ _` | '__/ _` |/ _ \ __|    / _ \ \ / / | |   | |   | |/ _ \| | | |/ _` |
+  ___) |  __/ |_     | | (_| | | | (_| |  __/ |_    / ___ \ V /  | |   | |___| | (_) | |_| | (_| |
+ |____/ \___|\__|    |_|\__,_|_|  \__, |\___|\__|  /_/   \_\_/  |___|   \____|_|\___/ \__,_|\__,_|
+                                  |___/                                                           
+
+'''
+
+titles_dnsprofile = '''
+
+  ____                     ____  _   _ ____     ____             __ _ _           
+ / ___|  ___ __ _ _ __    |  _ \| \ | / ___|   |  _ \ _ __ ___  / _(_) | ___  ___ 
+ \___ \ / __/ _` | '_ \   | | | |  \| \___ \   | |_) | '__/ _ \| |_| | |/ _ \/ __|
+  ___) | (_| (_| | | | |  | |_| | |\  |___) |  |  __/| | | (_) |  _| | |  __/\__ \.
+ |____/ \___\__,_|_| |_|  |____/|_| \_|____/   |_|   |_|  \___/|_| |_|_|\___||___/
+                                                                                  
+'''
+
+titles_ipamprofile = '''
+
+  ____                     ___ ____   _    __  __    ____             __ _ _           
+ / ___|  ___ __ _ _ __    |_ _|  _ \ / \  |  \/  |  |  _ \ _ __ ___  / _(_) | ___  ___ 
+ \___ \ / __/ _` | '_ \    | || |_) / _ \ | |\/| |  | |_) | '__/ _ \| |_| | |/ _ \/ __|
+  ___) | (_| (_| | | | |   | ||  __/ ___ \| |  | |  |  __/| | | (_) |  _| | |  __/\__ \.
+ |____/ \___\__,_|_| |_|  |___|_| /_/   \_\_|  |_|  |_|   |_|  \___/|_| |_|_|\___||___/
+                                                                                       
+'''
+
+titles_vrfcontext = '''
+
+  ____       _           _      _____                    _     __     ______  _____     ____            _            _   
+ / ___|  ___| | ___  ___| |_   |_   _|_ _ _ __ __ _  ___| |_   \ \   / /  _ \|  ___|   / ___|___  _ __ | |_ _____  _| |_ 
+ \___ \ / _ \ |/ _ \/ __| __|    | |/ _` | '__/ _` |/ _ \ __|   \ \ / /| |_) | |_     | |   / _ \| '_ \| __/ _ \ \/ / __|
+  ___) |  __/ |  __/ (__| |_     | | (_| | | | (_| |  __/ |_     \ V / |  _ <|  _|    | |__| (_) | | | | ||  __/>  <| |_ 
+ |____/ \___|_|\___|\___|\__|    |_|\__,_|_|  \__, |\___|\__|     \_/  |_| \_\_|       \____\___/|_| |_|\__\___/_/\_\\__|
+                                              |___/                                                                      
+
+'''
+titles_serviceenginegroup = '''
+
+  ____       _           _      _____                    _      ____                  _             _____             _             
+ / ___|  ___| | ___  ___| |_   |_   _|_ _ _ __ __ _  ___| |_   / ___|  ___ _ ____   _(_) ___ ___   | ____|_ __   __ _(_)_ __   ___  
+ \___ \ / _ \ |/ _ \/ __| __|    | |/ _` | '__/ _` |/ _ \ __|  \___ \ / _ \ '__\ \ / / |/ __/ _ \  |  _| | '_ \ / _` | | '_ \ / _ \ 
+  ___) |  __/ |  __/ (__| |_     | | (_| | | | (_| |  __/ |_    ___) |  __/ |   \ V /| | (_|  __/  | |___| | | | (_| | | | | |  __/ 
+ |____/_\___|_|\___|\___|\__|    |_|\__,_|_|  \__, |\___|\__|  |____/ \___|_|    \_/ |_|\___\___|  |_____|_| |_|\__, |_|_| |_|\___| 
+   / ___|_ __ ___  _   _ _ __                 |___/                                                             |___/               
+  | |  _| '__/ _ \| | | | '_ \                                                                                                      
+  | |_| | | | (_) | |_| | |_) |                                                                                                     
+   \____|_|  \___/ \__,_| .__/                                                                                                      
+                        |_|                                                                                                         
+
+'''
+
+titles_vs_display = '''
+
+  ____  _           _               __     ___      _               _    ____                  _               
+ |  _ \(_)___ _ __ | | __ _ _   _   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___ 
+ | | | | / __| '_ \| |/ _` | | | |   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|
+ | |_| | \__ \ |_) | | (_| | |_| |    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \\
+ |____/|_|___/ .__/|_|\__,_|\__, |     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/
+             |_|            |___/                                                                              
+'''
+
+titles_vs_selector = '''
+
+  ____       _           _     __     ___      _               _    ____                  _                
+ / ___|  ___| | ___  ___| |_   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___  
+ \___ \ / _ \ |/ _ \/ __| __|   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __| 
+  ___) |  __/ |  __/ (__| |_     \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \ 
+ |____/ \___|_|\___|\___|\__|     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/ 
+                                                                                                           
+
+'''
+
+titles_httppolicyset_selector = '''
+
+  ____       _           _      _   _ _____ _____ ____     ____       _ _               ____       _          
+ / ___|  ___| | ___  ___| |_   | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___    
+ \___ \ / _ \ |/ _ \/ __| __|  | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|   
+  ___) |  __/ |  __/ (__| |_   |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \   
+ |____/ \___|_|\___|\___|\__|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/   
+                                                                               |___/                          
+
+'''
+
+titles_httppolicyset_scanner = '''
+
+  ____                     _   _ _____ _____ ____     ____       _ _               ____       _                   
+ / ___|  ___ __ _ _ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___             
+ \___ \ / __/ _` | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|            
+  ___) | (_| (_| | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \            
+ |____/ \___\__,_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/            
+   __               ____             _                      ____          |___/   ____                            
+  / _| ___  _ __   |  _ \ ___   ___ | |___     ___  _ __   |  _ \ ___   ___ | |  / ___|_ __ ___  _   _ _ __  ___  
+ | |_ / _ \| '__|  | |_) / _ \ / _ \| / __|   / _ \| '__|  | |_) / _ \ / _ \| | | |  _| '__/ _ \| | | | '_ \/ __| 
+ |  _| (_) | |     |  __/ (_) | (_) | \__ \  | (_) | |     |  __/ (_) | (_) | | | |_| | | | (_) | |_| | |_) \__ \ 
+ |_|  \___/|_|     |_|   \___/ \___/|_|___/   \___/|_|     |_|   \___/ \___/|_|  \____|_|  \___/ \__,_| .__/|___/ 
+                                                                                                      |_|         
+
+'''
+
+titles_httppolicyset_migrate_pools = '''
+
+  __  __ _                 _          ____             _        _          _   _ _____ _____ ____     ____       _ _               ____       _           
+ |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |___   (_)_ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
+ | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| / __|  | | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
+ | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | \__ \  | | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
+ |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|___/  |_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
+           |___/                                                                                                          |___/                           
+
+'''
+
+titles_httppolicyset_migratepoolgroups = '''
+
+  __  __ _                 _          ____             _     ____                              _          _   _ _____ _____ ____     ____       _ _               ____       _           
+ |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |   / ___|_ __ ___  _   _ _ __  ___   (_)_ __    | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
+ | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| |  | |  _| '__/ _ \| | | | '_ \/ __|  | | '_ \   | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
+ | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | |  | |_| | | | (_) | |_| | |_) \__ \  | | | | |  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
+ |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|   \____|_|  \___/ \__,_| .__/|___/  |_|_| |_|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
+           |___/                                                                 |_|                                                                     |___/                           
+
+'''
+
+titles_httppolicyset_migrate = '''
+
+  __  __ _                 _          _   _ _____ _____ ____     ____       _ _               ____       _           
+ |  \/  (_) __ _ _ __ __ _| |_ ___   | | | |_   _|_   _|  _ \   |  _ \ ___ | (_) ___ _   _   / ___|  ___| |_ ___     
+ | |\/| | |/ _` | '__/ _` | __/ _ \  | |_| | | |   | | | |_) |  | |_) / _ \| | |/ __| | | |  \___ \ / _ \ __/ __|    
+ | |  | | | (_| | | | (_| | ||  __/  |  _  | | |   | | |  __/   |  __/ (_) | | | (__| |_| |   ___) |  __/ |_\__ \    
+ |_|  |_|_|\__, |_|  \__,_|\__\___|  |_| |_| |_|   |_| |_|      |_|   \___/|_|_|\___|\__, |  |____/ \___|\__|___/    
+           |___/                                                                     |___/                           
+
+'''
+
+titles_vs_migratepoolgroups = '''
+
+  __  __ _                 _          ____             _     ____                                     __   __     ___      _               _    ____                  _                  
+ |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |   / ___|_ __ ___  _   _ _ __  ___     ___  / _|  \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
+ | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| |  | |  _| '__/ _ \| | | | '_ \/ __|   / _ \| |_    \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
+ | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | |  | |_| | | | (_) | |_| | |_) \__ \  | (_) |  _|    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
+ |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|   \____|_|  \___/ \__,_| .__/|___/   \___/|_|       \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
+           |___/                                                                 |_|                                                                                                     
+
+'''
+
+titles_vs_migratepools = '''
+
+  __  __ _                 _          ____             _               __   __     ___      _               _    ____                  _                  
+ |  \/  (_) __ _ _ __ __ _| |_ ___   |  _ \ ___   ___ | |___     ___  / _|  \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
+ | |\/| | |/ _` | '__/ _` | __/ _ \  | |_) / _ \ / _ \| / __|   / _ \| |_    \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
+ | |  | | | (_| | | | (_| | ||  __/  |  __/ (_) | (_) | \__ \  | (_) |  _|    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
+ |_|  |_|_|\__, |_|  \__,_|\__\___|  |_|   \___/ \___/|_|___/   \___/|_|       \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
+           |___/                                                                                                                                          
+
+'''
+
+titles_vsvip_migrate = '''
+
+  __  __ _                 _         __     ______   __     _____ ____       
+ |  \/  (_) __ _ _ __ __ _| |_ ___   \ \   / / ___|  \ \   / /_ _|  _ \ ___  
+ | |\/| | |/ _` | '__/ _` | __/ _ \   \ \ / /\___ \   \ \ / / | || |_) / __| 
+ | |  | | | (_| | | | (_| | ||  __/    \ V /  ___) |   \ V /  | ||  __/\__ \ 
+ |_|  |_|_|\__, |_|  \__,_|\__\___|     \_/  |____/     \_/  |___|_|   |___/ 
+           |___/                                                             
+
+'''
+
+titles_vs_migrate = '''
+
+  __  __ _                 _         __     ___      _               _    ____                  _                  
+ |  \/  (_) __ _ _ __ __ _| |_ ___   \ \   / (_)_ __| |_ _   _  __ _| |  / ___|  ___ _ ____   _(_) ___ ___  ___    
+ | |\/| | |/ _` | '__/ _` | __/ _ \   \ \ / /| | '__| __| | | |/ _` | |  \___ \ / _ \ '__\ \ / / |/ __/ _ \/ __|   
+ | |  | | | (_| | | | (_| | ||  __/    \ V / | | |  | |_| |_| | (_| | |   ___) |  __/ |   \ V /| | (_|  __/\__ \   
+ |_|  |_|_|\__, |_|  \__,_|\__\___|     \_/  |_|_|   \__|\__,_|\__,_|_|  |____/ \___|_|    \_/ |_|\___\___||___/   
+           |___/                                                                                                   
+
+'''
+
+titles_logout = '''
+
+  _                            _                      _                   _                      _             
+ | |    ___   __ _  ___  _   _| |_     __ _ _ __   __| |    ___ _ __   __| |   ___  ___  ___ ___(_) ___  _ __  
+ | |   / _ \ / _` |/ _ \| | | | __|   / _` | '_ \ / _` |   / _ \ '_ \ / _` |  / __|/ _ \/ __/ __| |/ _ \| '_ \ 
+ | |__| (_) | (_| | (_) | |_| | |_   | (_| | | | | (_| |  |  __/ | | | (_| |  \__ \  __/\__ \__ \ | (_) | | | |
+ |_____\___/ \__, |\___/ \__,_|\__|   \__,_|_| |_|\__,_|   \___|_| |_|\__,_|  |___/\___||___/___/_|\___/|_| |_|
+             |___/                                                                                             
+
+'''
+
+title_remove_prefix = '''
+
+  ____                                   ____            __ _       
+ |  _ \ ___ _ __ ___   _____   _____    |  _ \ _ __ ___ / _(_)_  __ 
+ | |_) / _ \ '_ ` _ \ / _ \ \ / / _ \   | |_) | '__/ _ \ |_| \ \/ / 
+ |  _ <  __/ | | | | | (_) \ V /  __/   |  __/| | |  __/  _| |>  <  
+ |_| \_\___|_| |_| |_|\___/ \_/ \___|   |_|   |_|  \___|_| |_/_/\_\ 
+                                                                    
+
+'''
+
+titles_initiate_cleanup = '''
+
+  ___       _ _   _       _           ____ _                              
+ |_ _|_ __ (_) |_(_) __ _| |_ ___    / ___| | ___  __ _ _ __  _   _ _ __  
+  | || '_ \| | __| |/ _` | __/ _ \  | |   | |/ _ \/ _` | '_ \| | | | '_ \ 
+  | || | | | | |_| | (_| | ||  __/  | |___| |  __/ (_| | | | | |_| | |_) |
+ |___|_| |_|_|\__|_|\__,_|\__\___|   \____|_|\___|\__,_|_| |_|\__,_| .__/ 
+                                                                   |_|    
+
+'''
+
+titles_thanks = '''
+
+
+             _____  _                    _              __                              _                                               
+            |_   _|| |__    __ _  _ __  | | __ ___     / _|  ___   _ __     _   _  ___ (_) _ __    __ _                                 
+              | |  | '_ \  / _` || '_ \ | |/ // __|   | |_  / _ \ | '__|   | | | |/ __|| || '_ \  / _` |                                
+              | |  | | | || (_| || | | ||   < \__ \   |  _|| (_) || |      | |_| |\__ \| || | | || (_| |                                
+              |_|  |_| |_| \__,_||_| |_||_|\_\|___/   |_|   \___/ |_|       \__,_||___/|_||_| |_| \__, |                                
+                      __     __ __  __                                    _ __     __ ___         |___/                                 
+                      \ \   / /|  \/  |__      __ __ _  _ __  ___        / \\ \   / /|_ _|                                              
+                       \ \ / / | |\/| |\ \ /\ / // _` || '__|/ _ \      / _ \\ \ / /  | |                                               
+                        \ V /  | |  | | \ V  V /| (_| || |  |  __/     / ___ \\ V /   | |                                               
+                         \_/   |_|  |_|  \_/\_/  \__,_||_|   \___|    /_/   \_\\_/   |___|                                              
+ __     __ _        _                  _      ____                      _                 __  __  _                     _               
+ \ \   / /(_) _ __ | |_  _   _   __ _ | |    / ___|   ___  _ __ __   __(_)  ___  ___     |  \/  |(_)  __ _  _ __  __ _ | |_  ___   _ __ 
+  \ \ / / | || '__|| __|| | | | / _` || |    \___ \  / _ \| '__|\ \ / /| | / __|/ _ \    | |\/| || | / _` || '__|/ _` || __|/ _ \ | '__|
+   \ V /  | || |   | |_ | |_| || (_| || |     ___) ||  __/| |    \ V / | || (__|  __/    | |  | || || (_| || |  | (_| || |_| (_) || |   
+    \_/   |_||_|    \__| \__,_| \__,_||_|    |____/  \___||_|     \_/  |_| \___|\___|    |_|  |_||_| \__, ||_|   \__,_| \__|\___/ |_|   
+                                                                                                     |___/                              
+  
+
+'''
